Update react package documentation
