--- packages/sdk-react/docs/documentation.md
@@ -1,18 +1,0 @@
-# Documentation
-
-### Context
-
-- [FusionAuthProvider](context.md#fusionauthprovider)
-- [FusionAuthContext](context.md#fusionauthcontext)
-
-### Components
-
-- [FusionAuthLoginButton](components.md#fusionauthloginbutton)
-- [FusionAuthLogoutButton](components.md#fusionauthlogoutbutton)
-- [FusionAuthRegisterButton](components.md#fusionauthregisterbutton)
-- [RequireAuth](components.md#requireauth)
-
-### Functions
-
-- [useFusionAuth](functions.md#usefusionauth)
-- [withFusionAuth](functions.md#withfusionauth)
