+++ packages/sdk-react/docs/modules.md
@@ -1,0 +1,11 @@
+[@fusionauth/react-sdk](README.md) / Modules
+
+# @fusionauth/react-sdk
+
+## Table of contents
+
+### Modules
+
+- [FusionAuthProvider](modules/FusionAuthProvider.md)
+- [FusionAuthProviderConfig](modules/FusionAuthProviderConfig.md)
+- [FusionAuthProviderContext](modules/FusionAuthProviderContext.md)
