+++ packages/sdk-react/docs/modules/FusionAuthProviderConfig.md
@@ -1,0 +1,9 @@
+[@fusionauth/react-sdk](../README.md) / [Modules](../modules.md) / FusionAuthProviderConfig
+
+# Module: FusionAuthProviderConfig
+
+## Table of contents
+
+### Interfaces
+
+- [FusionAuthProviderConfig](../interfaces/FusionAuthProviderConfig.FusionAuthProviderConfig.md)
