+++ packages/sdk-react/src/components/ui/FusionAuthRegisterButton/index.tsx
@@ -9,6 +9,7 @@ interface Props {
   className?: string;
 }
 
+/** Calls the `startRegister` method from `useFusionAuth`. */
 export const FusionAuthRegisterButton: FC<Props> = ({
   state,
   text,
