+++ packages/sdk-react/tsconfig.json
@@ -12,5 +12,14 @@
     "outDir": "dist",
     "jsx": "react-jsx"
   },
-  "include": ["src", "vite.config.ts"]
+  "include": ["src", "vite.config.ts"],
+  "typedocOptions": {
+    "plugin": ["typedoc-plugin-markdown"],
+    "entryPoints": [
+      "src/components/providers/FusionAuthProvider.tsx",
+      "src/components/providers/FusionAuthProviderConfig.ts",
+      "src/components/providers/FusionAuthProviderContext.ts"
+    ],
+    "out": "docs"
+  }
 }
